Added automatic dependency resolution of header files
This uses the -MMD flag for compiling.

When compiling with this flag, a .d file with the same name as the argument of
the -o option (but with .o replaced by .d) is created.

An include directive is added to the makefile (by adding it at the end of
makefile.in) to take in the info for those includes.

This guarantees that all files that need to be recompiled following a change to
a header file will be recompiled.

see make.mad-scientist.net/papers/advanced-auto-dependency-generation/